norix-image-generator-cloud/requirements.txt
streamlit
openai
google-genai
anthropic
python-dotenv
pillow
pandas
